+++ app/Country.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Models;
+namespace App;
 
 use App\Brewery;
 use Illuminate\Database\Eloquent\Model;
